RENAME: `Tooltip` to `TextTooltip`, `TooltipBase` to `Tooltip` [complete]
