+++ src/components/Tooltips/__tests__/TextTooltip.test.tsx
@@ -1,7 +1,7 @@
 import { render, fireEvent } from "@tests/test-utils";
 // test utils
 
-import Tooltip from "@components/Tooltips/Tooltip";
+import TextTooltip from "@components/Tooltips/TextTooltip";
 // components
 
 import { initializeObserver } from "@components/Tooltips/__tests__/__mock__/mockResizeObserver";
@@ -18,9 +18,9 @@ describe("Tooltip", () => {
 
   it("툴팁은 하위 요소 렌더링에 영향을 주지 않는다.", () => {
     const { getByRole } = render(
-      <Tooltip text={TOOLTIP_TEXT}>
+      <TextTooltip text={TOOLTIP_TEXT}>
         <button type="button">{BUTTON_CONTENT}</button>
-      </Tooltip>
+      </TextTooltip>
     );
 
     const button = getByRole("button");
@@ -30,9 +30,9 @@ describe("Tooltip", () => {
 
   it("툴팁은 기본적으로 렌더링되지 않으며, 하위 요소에 `mouseover` 시 툴팁이 body 아래에 `text`를 포함하여 렌더링된 후 일정 시간이 지나면 사라진다.", async () => {
     const { getByRole, queryByRole } = render(
-      <Tooltip text={TOOLTIP_TEXT}>
+      <TextTooltip text={TOOLTIP_TEXT}>
         <button type="button">{BUTTON_CONTENT}</button>
-      </Tooltip>
+      </TextTooltip>
     );
 
     const button = getByRole("button");
@@ -52,9 +52,9 @@ describe("Tooltip", () => {
 
   it("툴팁의 `z-index` 기본값은 `1,000`이다.", async () => {
     const { getByRole } = render(
-      <Tooltip text={TOOLTIP_TEXT}>
+      <TextTooltip text={TOOLTIP_TEXT}>
         <button type="button">{BUTTON_CONTENT}</button>
-      </Tooltip>
+      </TextTooltip>
     );
     const button = getByRole("button");
 
@@ -67,9 +67,9 @@ describe("Tooltip", () => {
 
   it("툴팁의 `z-index`는 설정할 수 있다.", async () => {
     const { getByRole } = render(
-      <Tooltip text={TOOLTIP_TEXT} zIndex={1500}>
+      <TextTooltip text={TOOLTIP_TEXT} zIndex={1500}>
         <button type="button">{BUTTON_CONTENT}</button>
-      </Tooltip>
+      </TextTooltip>
     );
     const button = getByRole("button");
