+++ src/components/Tooltips/styles/Tooltip.style.ts
@@ -1,11 +1,11 @@
-import type { TooltipProps } from "@components/Tooltips/Tooltip";
+import type { TextTooltipProps } from "@components/Tooltips/TextTooltip";
 // interfaces
 
 import styled, { css } from "@themes/styled";
 // styles
 
 // eslint-disable-next-line import/prefer-default-export
-export const StyleTooltipText = styled.p<Pick<TooltipProps, "textSize">>`
+export const StyleTooltipText = styled.p<Pick<TextTooltipProps, "textSize">>`
   color: #fff;
   ${({ theme }) => theme.font.kopub.contents5}
